+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 #### Fixed
 - Logs produced by language server in WebAssembly mode are not visible
+- CCW is assigned incorrect symbol attributes in the lookahead mode
 
 ## [1.18.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.17.0...1.18.0) (2025-06-20)